pyproject.toml
[tool.black]
line-length = 110
target-version = ["py38"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.8
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict = true
plugins = ["numpy.typing.mypy_plugin"]

[tool.poetry]
name = "multi-repo-automation"
version = "0.0.0"
description = "Library for automation updates on multiple repositories."
readme = "README.md"
keywords = ["pre-commit"]
license = "BSD-2-Clause"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed",
]
authors = ["Stéphane Brunner <stephane.brunner@gmail.com>"]
repository = "https://github.com/sbrunner/multi-repo-automation"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
PyYAML = "6.0"
requests = "2.28.2"
"ruamel.yaml" = "0.17.21"
identify = "2.5.22"
c2cciutils = { version = "1.5.2", optional = true }
tomlkit = "0.11.7"
configupdater = "3.1.1"

[tool.poetry.extras]
update_stabilization_branches = ["c2cciutils"]

[tool.poetry.group.dev.dependencies]
prospector = { extras = ["with_bandit", "with_mypy", "with_pyroma"], version = "1.9.0" }
types-PyYAML = "6.0.12.9"
types-requests = "2.28.11.17"

[tool.poetry.scripts]
mra-update-stabilization-branches = 'multi_repo_automation:update_stabilization_branches_main'

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning", "poetry-plugin-tweak-dependencies-version"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
pattern = "^(?P<base>\\d+(\\.\\d+)*)"
format-jinja = """
{%- if env.get("VERSION_TYPE") == "version_branch" -%}
{{serialize_pep440(bump_version(base, 1 if env.get("IS_MASTER") == "TRUE" else 2), dev=distance)}}
{%- elif distance == 0 -%}
{{serialize_pep440(base)}}
{%- else -%}
{{serialize_pep440(bump_version(base), dev=distance)}}
{%- endif -%}
"""

[tool.poetry-plugin-tweak-dependencies-version]
default = "present"
